requirements.txt
fastapi
uvicorn
python-dotenv
supabase
transformers
torch
tf-keras
google-generativeai
